requirements.txt
pycryptodomex
pywin32
pycryptodome